+++ spec/assets/sample1/f1.js
@@ -1,0 +1,1 @@
+var a=1,b=[1,2,3];b.forEach(function(n){return n+1})