+++ spec/assets/sample1/inner/f2.js
@@ -1,0 +1,1 @@
+function f(a,b,c){return [a,b,c].map(function(n){n*2})}